requirements.txt
flask~=2.3.3
Flask-SQLAlchemy
flask-login
werkzeug~=2.3.7
SQLAlchemy~=2.0.20
WTForms~=3.0.1
python-dotenv~=1.0.0
beautifulsoup4~=4.12.2
flask_talisman
flask_wtf
email_validator
